ci-2

+++ ci/run-ci-gpu-tests.bash
@@ -9,25 +9,27 @@ assert_env() {
     fi
 }
 
-echo "::group::Environment"
 assert_env PUNICA_CI_CACHE_DIR
 assert_env PUNICA_CI_TORCH_VERSION
 assert_env PUNICA_CI_CUDA_MAJOR
 assert_env PUNICA_CI_CUDA_MINOR
 PROJECT_ROOT="$(cd "$(dirname "$0")/.." && pwd)"
 export CONDA_pkgs_dirs="$PUNICA_CI_CACHE_DIR/conda-pkgs"
-PIP_CACHE_DIR="$PUNICA_CI_CACHE_DIR/pip"
-mkdir -p "$CONDA_pkgs_dirs" "$PIP_CACHE_DIR"
+export XDG_CACHE_HOME="$PUNICA_CI_CACHE_DIR/xdg-cache"
+mkdir -p "$CONDA_pkgs_dirs" "$XDG_CACHE_HOME"
+export HOME=/tmp/home
+mkdir -p $HOME
 nvidia-smi
-echo "::endgroup::"
 
 
 echo "::group::Install Mamba and Python"
-wget -O "$PUNICA_CI_CACHE_DIR/Miniforge3.sh" "https://github.com/conda-forge/miniforge/releases/latest/download/Miniforge3-$(uname)-$(uname -m).sh"
-"$PUNICA_CI_CACHE_DIR/Miniforge3.sh" -b -p "/tmp/conda"
-source "/tmp/conda/etc/profile.d/conda.sh"
-source "/tmp/conda/etc/profile.d/mamba.sh"
-mamba create -y -n punica-ci python=3.10
+if [ ! -f "$PUNICA_CI_CACHE_DIR/Miniforge3.sh" ]; then
+    wget -O "$PUNICA_CI_CACHE_DIR/Miniforge3.sh" "https://github.com/conda-forge/miniforge/releases/latest/download/Miniforge3-$(uname)-$(uname -m).sh"
+fi
+bash "$PUNICA_CI_CACHE_DIR/Miniforge3.sh" -b -p "$HOME/conda"
+source "$HOME/conda/etc/profile.d/conda.sh"
+source "$HOME/conda/etc/profile.d/mamba.sh"
+mamba create -y -n punica-ci python=3.10.13
 mamba activate punica-ci
 echo "::endgroup::"
 
@@ -40,7 +42,8 @@ echo "::endgroup::"
 echo "::group::Install Punica"
 cd "$PROJECT_ROOT"
 pip install ninja numpy
-pip install -v --no-build-isolation .[dev]
+# pip install -v --no-build-isolation .[dev]
+pip install -v -e .[dev]
 echo "::endgroup::"
